normi for most
