+++ src/cd.c
@@ -6,56 +6,11 @@
 /*   By: sionow <sionow@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/03 23:05:48 by sionow            #+#    #+#             */
-/*   Updated: 2025/08/23 18:59:29 by sionow           ###   ########.fr       */
+/*   Updated: 2025/08/26 17:18:08 by sionow           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-int	ft_cd_error(char *path, char *to_free)
-{
-	write(2, "cd: ", 4);
-	write(2, path, ft_strlen(path));
-	write(2, ": No such file or directory\n", 28);
-	if (to_free != NULL)
-		free(to_free);
-	return (1);
-}
-
-char	*ft_strjoinslash(char *s1, char *s2)
-{
-	int		s1m;
-	int		s2m;
-	int		len;
-	char	*finals;
-
-	s1m = 0;
-	s2m = 0;
-	len = ft_strlen(s1) + ft_strlen(s2);
-	finals = malloc (sizeof(char) * (len + 2));
-	if (!finals)
-		return (NULL);
-	while (s1[s1m])
-	{
-		finals[s1m] = s1[s1m];
-		s1m++;
-	}
-	finals[s1m++] = '/';
-	while (s2[s2m])
-	{
-		finals[s1m++] = s2[s2m++];
-	}
-	finals[s1m] = '\0';
-	return (finals);
-}
-
-char	*fill_path(char *str)
-{
-	char	cwd[PATH_MAX];
-
-	getcwd(cwd, sizeof(cwd));
-	return (ft_strjoinslash(cwd, &str[2]));
-}
 
 int	ft_path_extra(char *str, t_pipeline *pl)
 {
@@ -99,13 +54,6 @@ int	ft_paths(char *str)
 			return (ft_cd_error(str, NULL));
 		return (0);
 	}
-	return (1);
-}
-
-int	error_msg_cd()
-{
-	write (2, "cd: ", 4);
-	write (2, "OLDPWD not set\n", 16);
 	return (1);
 }
