+++ src/parent.c
@@ -6,7 +6,7 @@
 /*   By: sionow <sionow@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/14 22:31:38 by sionow            #+#    #+#             */
-/*   Updated: 2025/08/23 23:21:17 by sionow           ###   ########.fr       */
+/*   Updated: 2025/08/26 18:58:57 by sionow           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,4 +34,3 @@ int	adoption_center(t_cmd *cmds)
 		return (0);
 	return (1);
 }
-
