+++ src/unset.c
@@ -6,13 +6,13 @@
 /*   By: sionow <sionow@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/18 23:59:53 by sionow            #+#    #+#             */
-/*   Updated: 2025/08/26 16:49:33 by sionow           ###   ########.fr       */
+/*   Updated: 2025/08/26 18:36:58 by sionow           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int unset_head(char *str, t_pipeline *pl)
+int	unset_head(char *str, t_pipeline *pl)
 {
 	t_env_var	*finger;
 	t_env_var	*temp;
@@ -33,14 +33,14 @@ int unset_head(char *str, t_pipeline *pl)
 
 void	unset_last(char *str, t_pipeline *pl)
 {
-		t_env_var	*finger;
+	t_env_var	*finger;
 	t_env_var	*temp;
 
 	finger = pl->env;
 	temp = finger;
 	while (finger->next && finger->next->next)
 		finger = finger->next;
-	if (finger->next && finger->next->next == NULL 
+	if (finger->next && finger->next->next == NULL
 		&& ft_strncmp(str, finger->next->fullstring, ft_strlen(str)) == 0
 		&& finger->next->fullstring[ft_strlen(str)] == '=')
 	{
@@ -79,7 +79,7 @@ void	unset_env(char *str, t_pipeline *pl)
 	return ;
 }
 
-int ft_unset(int argc, char **argv, t_pipeline *pl)
+int	ft_unset(int argc, char **argv, t_pipeline *pl)
 {
 	int	i;
